Moved Autocompletion to new tab and moved Custom tab editor to EntryEditorTab, reordered tabs
