+++ src/main/java/org/jabref/gui/preferences/entryeditor/EntryEditorTab.fxml
@@ -1,13 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.RadioButton?>
-<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.TextArea?>
 <?import javafx.scene.control.ToggleGroup?>
+<?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
+<?import org.jabref.gui.icon.JabRefIconView?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.entryeditor.EntryEditorTab">
@@ -30,33 +32,22 @@
     <CheckBox fx:id="enableValidation" text="%Show validation messages"/>
     <CheckBox fx:id="allowIntegerEdition" text="%Allow integers in 'edition' field in BibTeX mode"/>
 
-    <Label styleClass="sectionHeader" text="%Autocompletion"/>
-    <CheckBox fx:id="enableAutoComplete" text="%Use autocompletion"/>
-    <VBox spacing="10.0">
-        <HBox spacing="4.0">
-            <Label text="%Affected fields" disable="${!enableAutoComplete.selected}"/>
-            <TextField fx:id="autoCompleteFields" HBox.hgrow="ALWAYS" disable="${!enableAutoComplete.selected}">
-                <HBox.margin>
-                    <Insets top="-4.0"/>
-                </HBox.margin>
-            </TextField>
-        </HBox>
-        <Label text="%Name format" disable="${!enableAutoComplete.selected}"/>
-        <RadioButton fx:id="autoCompleteFirstLast" text="%Autocomplete names in 'Firstname Lastname' format only"
-                     toggleGroup="$nameFormat" disable="${!enableAutoComplete.selected}"/>
-        <RadioButton fx:id="autoCompleteLastFirst" text="%Autocomplete names in 'Lastname, Firstname' format only"
-                     toggleGroup="$nameFormat" disable="${!enableAutoComplete.selected}"/>
-        <RadioButton fx:id="autoCompleteBoth" text="%Autocomplete names in both formats"
-                     toggleGroup="$nameFormat" disable="${!enableAutoComplete.selected}"/>
-        <Label text="%First names" disable="${!enableAutoComplete.selected}"/>
-        <RadioButton fx:id="firstNameModeAbbreviated" text="%Use abbreviated firstname whenever possible"
-                     toggleGroup="$firstNames" disable="${!enableAutoComplete.selected}"/>
-        <RadioButton fx:id="firstNameModeFull" text="%Use full firstname whenever possible"
-                     toggleGroup="$firstNames" disable="${!enableAutoComplete.selected}"/>
-        <RadioButton fx:id="firstNameModeBoth" text="%Use abbreviated and full firstname"
-                     toggleGroup="$firstNames" disable="${!enableAutoComplete.selected}"/>
-        <padding>
-            <Insets left="20.0"/>
-        </padding>
-    </VBox>
+    <Label styleClass="sectionHeader" text="%Custom editor tabs"/>
+    <HBox>
+        <Label text="%Create custom fields for each BibTeX entry"/>
+        <HBox HBox.hgrow="ALWAYS"/>
+        <Button fx:id="generalFieldsHelp"/>
+        <Button prefHeight="20.0" prefWidth="20.0" styleClass="icon-button,narrow" onAction="#resetToDefaults">
+            <graphic>
+                <JabRefIconView glyph="REFRESH"/>
+            </graphic>
+            <tooltip>
+                <Tooltip text="%Reset to default"/>
+            </tooltip>
+        </Button>
+    </HBox>
+
+    <TextArea fx:id="fieldsTextArea" minHeight="200.0"/>
+
+    <Label text="%Format: Tab:field;field;... (e.g. General:url;pdf;note...)"/>
 </fx:root>
