+++ src/main/java/org/jabref/gui/preferences/entryeditor/EntryEditorTab.java
@@ -1,12 +1,17 @@
 package org.jabref.gui.preferences.entryeditor;
 
 import javafx.fxml.FXML;
+import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
-import javafx.scene.control.RadioButton;
-import javafx.scene.control.TextField;
+import javafx.scene.control.TextArea;
 
+import org.jabref.gui.Globals;
+import org.jabref.gui.actions.ActionFactory;
+import org.jabref.gui.actions.StandardActions;
+import org.jabref.gui.help.HelpAction;
 import org.jabref.gui.preferences.AbstractPreferenceTabView;
 import org.jabref.gui.preferences.PreferencesTab;
+import org.jabref.logic.help.HelpFile;
 import org.jabref.logic.l10n.Localization;
 
 import com.airhacks.afterburner.views.ViewLoader;
@@ -20,14 +25,9 @@ public class EntryEditorTab extends AbstractPreferenceTabView<EntryEditorTabView
     @FXML private CheckBox enableLatexCitationsTab;
     @FXML private CheckBox enableValidation;
     @FXML private CheckBox allowIntegerEdition;
-    @FXML private CheckBox enableAutoComplete;
-    @FXML private TextField autoCompleteFields;
-    @FXML private RadioButton autoCompleteFirstLast;
-    @FXML private RadioButton autoCompleteLastFirst;
-    @FXML private RadioButton autoCompleteBoth;
-    @FXML private RadioButton firstNameModeAbbreviated;
-    @FXML private RadioButton firstNameModeFull;
-    @FXML private RadioButton firstNameModeBoth;
+
+    @FXML private Button generalFieldsHelp;
+    @FXML private TextArea fieldsTextArea;
 
     public EntryEditorTab() {
         ViewLoader.view(this)
@@ -41,7 +41,7 @@ public class EntryEditorTab extends AbstractPreferenceTabView<EntryEditorTabView
     }
 
     public void initialize() {
-        this.viewModel = new EntryEditorTabViewModel(preferencesService.getEntryEditorPreferences(), preferencesService.getAutoCompletePreferences());
+        this.viewModel = new EntryEditorTabViewModel(dialogService, preferencesService);
 
         openOnNewEntry.selectedProperty().bindBidirectional(viewModel.openOnNewEntryProperty());
         defaultSource.selectedProperty().bindBidirectional(viewModel.defaultSourceProperty());
@@ -50,13 +50,15 @@ public class EntryEditorTab extends AbstractPreferenceTabView<EntryEditorTabView
         enableLatexCitationsTab.selectedProperty().bindBidirectional(viewModel.enableLatexCitationsTabProperty());
         enableValidation.selectedProperty().bindBidirectional(viewModel.enableValidationProperty());
         allowIntegerEdition.selectedProperty().bindBidirectional(viewModel.allowIntegerEditionProperty());
-        enableAutoComplete.selectedProperty().bindBidirectional(viewModel.enableAutoCompleteProperty());
-        autoCompleteFields.textProperty().bindBidirectional(viewModel.autoCompleteFieldsProperty());
-        autoCompleteFirstLast.selectedProperty().bindBidirectional(viewModel.autoCompleteFirstLastProperty());
-        autoCompleteLastFirst.selectedProperty().bindBidirectional(viewModel.autoCompleteLastFirstProperty());
-        autoCompleteBoth.selectedProperty().bindBidirectional(viewModel.autoCompleteBothProperty());
-        firstNameModeAbbreviated.selectedProperty().bindBidirectional(viewModel.firstNameModeAbbreviatedProperty());
-        firstNameModeFull.selectedProperty().bindBidirectional(viewModel.firstNameModeFullProperty());
-        firstNameModeBoth.selectedProperty().bindBidirectional(viewModel.firstNameModeBothProperty());
+
+        fieldsTextArea.textProperty().bindBidirectional(viewModel.fieldsProperty());
+
+        ActionFactory actionFactory = new ActionFactory(Globals.getKeyPrefs());
+        actionFactory.configureIconButton(StandardActions.HELP, new HelpAction(HelpFile.GENERAL_FIELDS, dialogService), generalFieldsHelp);
+    }
+
+    @FXML
+    void resetToDefaults() {
+        viewModel.resetToDefaults();
     }
 }
