+++ src/main/java/org/jabref/gui/preferences/entryeditor/EntryEditorTabViewModel.java
@@ -1,19 +1,22 @@
 package org.jabref.gui.preferences.entryeditor;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Set;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
-import org.jabref.gui.autocompleter.AutoCompleteFirstNameMode;
-import org.jabref.gui.autocompleter.AutoCompletePreferences;
+import org.jabref.gui.DialogService;
 import org.jabref.gui.entryeditor.EntryEditorPreferences;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
+import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.FieldFactory;
+import org.jabref.preferences.PreferencesService;
 
 public class EntryEditorTabViewModel implements PreferenceTabViewModel {
 
@@ -24,23 +27,17 @@ public class EntryEditorTabViewModel implements PreferenceTabViewModel {
     private final BooleanProperty enableLatexCitationsTabProperty = new SimpleBooleanProperty();
     private final BooleanProperty enableValidationProperty = new SimpleBooleanProperty();
     private final BooleanProperty allowIntegerEditionProperty = new SimpleBooleanProperty();
-    private final BooleanProperty enableAutoCompleteProperty = new SimpleBooleanProperty();
-    private final StringProperty autoCompleteFieldsProperty = new SimpleStringProperty();
-    private final BooleanProperty autoCompleteFirstLastProperty = new SimpleBooleanProperty();
-    private final BooleanProperty autoCompleteLastFirstProperty = new SimpleBooleanProperty();
-    private final BooleanProperty autoCompleteBothProperty = new SimpleBooleanProperty();
-    private final BooleanProperty firstNameModeAbbreviatedProperty = new SimpleBooleanProperty();
-    private final BooleanProperty firstNameModeFullProperty = new SimpleBooleanProperty();
-    private final BooleanProperty firstNameModeBothProperty = new SimpleBooleanProperty();
 
+    private final StringProperty fieldsProperty = new SimpleStringProperty();
+
+    private final DialogService dialogService;
+    private final PreferencesService preferencesService;
     private final EntryEditorPreferences entryEditorPreferences;
-    private final AutoCompletePreferences autoCompletePreferences;
 
-    private final List<String> restartWarnings = new ArrayList<>();
-
-    public EntryEditorTabViewModel(EntryEditorPreferences entryEditorPreferences, AutoCompletePreferences autoCompletePreferences) {
-        this.entryEditorPreferences = entryEditorPreferences;
-        this.autoCompletePreferences = autoCompletePreferences;
+    public EntryEditorTabViewModel(DialogService dialogService, PreferencesService preferencesService) {
+        this.dialogService = dialogService;
+        this.preferencesService = preferencesService;
+        this.entryEditorPreferences = preferencesService.getEntryEditorPreferences();
     }
 
     @Override
@@ -56,22 +53,24 @@ public class EntryEditorTabViewModel implements PreferenceTabViewModel {
         enableValidationProperty.setValue(entryEditorPreferences.shouldEnableValidation());
         allowIntegerEditionProperty.setValue(entryEditorPreferences.shouldAllowIntegerEditionBibtex());
 
-        enableAutoCompleteProperty.setValue(autoCompletePreferences.shouldAutoComplete());
-        autoCompleteFieldsProperty.setValue(autoCompletePreferences.getCompleteNamesAsString());
+        setFields(entryEditorPreferences.getEntryEditorTabList());
+    }
 
-        if (autoCompletePreferences.getNameFormat() == AutoCompletePreferences.NameFormat.FIRST_LAST) {
-            autoCompleteFirstLastProperty.setValue(true);
-        } else if (autoCompletePreferences.getNameFormat() == AutoCompletePreferences.NameFormat.LAST_FIRST) {
-            autoCompleteLastFirstProperty.setValue(true);
-        } else {
-            autoCompleteBothProperty.setValue(true);
-        }
+    public void resetToDefaults() {
+        setFields(preferencesService.getDefaultTabNamesAndFields());
+    }
 
-        switch (autoCompletePreferences.getFirstNameMode()) {
-            case ONLY_ABBREVIATED -> firstNameModeAbbreviatedProperty.setValue(true);
-            case ONLY_FULL -> firstNameModeFullProperty.setValue(true);
-            default -> firstNameModeBothProperty.setValue(true);
+    private void setFields(Map<String, Set<Field>> tabNamesAndFields) {
+        StringBuilder sb = new StringBuilder();
+
+        // Fill with customized vars
+        for (Map.Entry<String, Set<Field>> tab : tabNamesAndFields.entrySet()) {
+            sb.append(tab.getKey());
+            sb.append(':');
+            sb.append(FieldFactory.serializeFieldsList(tab.getValue()));
+            sb.append('\n');
         }
+        fieldsProperty.set(sb.toString());
     }
 
     @Override
@@ -86,39 +85,35 @@ public class EntryEditorTabViewModel implements PreferenceTabViewModel {
         entryEditorPreferences.setAllowIntegerEditionBibtex(allowIntegerEditionProperty.getValue());
         // entryEditorPreferences.setDividerPosition();
 
-        autoCompletePreferences.setAutoComplete(enableAutoCompleteProperty.getValue());
+        Map<String, Set<Field>> customTabsMap = new LinkedHashMap<>();
+        String[] lines = fieldsProperty.get().split("\n");
 
-        if (autoCompleteBothProperty.getValue()) {
-            autoCompletePreferences.setNameFormat(AutoCompletePreferences.NameFormat.BOTH);
-        } else if (autoCompleteFirstLastProperty.getValue()) {
-            autoCompletePreferences.setNameFormat(AutoCompletePreferences.NameFormat.FIRST_LAST);
-        } else if (autoCompleteLastFirstProperty.getValue()) {
-            autoCompletePreferences.setNameFormat(AutoCompletePreferences.NameFormat.LAST_FIRST);
-        }
-
-        if (firstNameModeBothProperty.getValue()) {
-            autoCompletePreferences.setFirstNameMode(AutoCompleteFirstNameMode.BOTH);
-        } else if (firstNameModeAbbreviatedProperty.getValue()) {
-            autoCompletePreferences.setFirstNameMode(AutoCompleteFirstNameMode.ONLY_ABBREVIATED);
-        } else if (firstNameModeFullProperty.getValue()) {
-            autoCompletePreferences.setFirstNameMode(AutoCompleteFirstNameMode.ONLY_FULL);
-        }
-
-        if (autoCompletePreferences.shouldAutoComplete() != enableAutoCompleteProperty.getValue()) {
-            if (enableAutoCompleteProperty.getValue()) {
-                restartWarnings.add(Localization.lang("Auto complete enabled."));
-            } else {
-                restartWarnings.add(Localization.lang("Auto complete disabled."));
+        for (String line : lines) {
+            String[] parts = line.split(":");
+            if (parts.length != 2) {
+                dialogService.showInformationDialogAndWait(
+                        Localization.lang("Error"),
+                        Localization.lang("Each line must be of the following form: 'tab:field1;field2;...;fieldN'."));
+                return;
             }
+
+            // Use literal string of unwanted characters specified below as opposed to exporting characters
+            // from preferences because the list of allowable characters in this particular differs
+            // i.e. ';' character is allowed in this window, but it's on the list of unwanted chars in preferences
+            String unwantedChars = "#{}()~,^&-\"'`ʹ\\";
+            String testString = CitationKeyGenerator.cleanKey(parts[1], unwantedChars);
+            if (!testString.equals(parts[1])) {
+                dialogService.showInformationDialogAndWait(
+                        Localization.lang("Error"),
+                        Localization.lang("Field names are not allowed to contain white spaces or certain characters (%0).",
+                                "# { } ( ) ~ , ^ & - \" ' ` ʹ \\"));
+                return;
+            }
+
+            customTabsMap.put(parts[0], FieldFactory.parseFieldList(parts[1]));
         }
 
-        autoCompletePreferences.getCompleteFields().clear();
-        autoCompletePreferences.getCompleteFields().addAll(FieldFactory.parseFieldList(autoCompleteFieldsProperty.getValue()));
-    }
-
-    @Override
-    public List<String> getRestartWarnings() {
-        return restartWarnings;
+        entryEditorPreferences.setEntryEditorTabList(customTabsMap);
     }
 
     public BooleanProperty openOnNewEntryProperty() {
@@ -149,35 +144,7 @@ public class EntryEditorTabViewModel implements PreferenceTabViewModel {
         return this.allowIntegerEditionProperty;
     }
 
-    public BooleanProperty enableAutoCompleteProperty() {
-        return enableAutoCompleteProperty;
-    }
-
-    public StringProperty autoCompleteFieldsProperty() {
-        return autoCompleteFieldsProperty;
-    }
-
-    public BooleanProperty autoCompleteFirstLastProperty() {
-        return autoCompleteFirstLastProperty;
-    }
-
-    public BooleanProperty autoCompleteLastFirstProperty() {
-        return autoCompleteLastFirstProperty;
-    }
-
-    public BooleanProperty autoCompleteBothProperty() {
-        return autoCompleteBothProperty;
-    }
-
-    public BooleanProperty firstNameModeAbbreviatedProperty() {
-        return firstNameModeAbbreviatedProperty;
-    }
-
-    public BooleanProperty firstNameModeFullProperty() {
-        return firstNameModeFullProperty;
-    }
-
-    public BooleanProperty firstNameModeBothProperty() {
-        return firstNameModeBothProperty;
+    public StringProperty fieldsProperty() {
+        return fieldsProperty;
     }
 }
