--- src/main/java/org/jabref/gui/preferences/entryeditortabs/EntryEditorTabsTab.fxml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.control.Tooltip?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import org.jabref.gui.icon.JabRefIconView?>
-<fx:root spacing="10.0" type="VBox"
-         xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-         fx:controller="org.jabref.gui.preferences.entryeditortabs.EntryEditorTabsTab">
-
-    <Label styleClass="titleHeader" text="%Custom editor tabs"/>
-
-    <HBox>
-        <Label text="%Create custom fields for each BibTeX entry"/>
-        <HBox HBox.hgrow="ALWAYS"/>
-        <Button fx:id="generalFieldsHelp"/>
-        <Button prefHeight="20.0" prefWidth="20.0" styleClass="icon-button,narrow" onAction="#resetToDefaults">
-            <graphic>
-                <JabRefIconView glyph="REFRESH"/>
-            </graphic>
-            <tooltip>
-                <Tooltip text="%Reset to default"/>
-            </tooltip>
-        </Button>
-    </HBox>
-
-    <TextArea fx:id="fieldsTextArea" minHeight="200.0"/>
-
-    <Label text="%Format: Tab:field;field;... (e.g. General:url;pdf;note...)"/>
-</fx:root>
