--- src/main/java/org/jabref/gui/preferences/entryeditortabs/EntryEditorTabsTabViewModel.java
@@ -1,91 +1,0 @@
-package org.jabref.gui.preferences.entryeditortabs;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Set;
-
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
-
-import org.jabref.gui.DialogService;
-import org.jabref.gui.entryeditor.EntryEditorPreferences;
-import org.jabref.gui.preferences.PreferenceTabViewModel;
-import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
-import org.jabref.logic.l10n.Localization;
-import org.jabref.model.entry.field.Field;
-import org.jabref.model.entry.field.FieldFactory;
-import org.jabref.preferences.PreferencesService;
-
-public class EntryEditorTabsTabViewModel implements PreferenceTabViewModel {
-
-    private final StringProperty fieldsProperty = new SimpleStringProperty();
-
-    private final DialogService dialogService;
-    private final PreferencesService preferences;
-    private final EntryEditorPreferences entryEditorPreferences;
-
-    public EntryEditorTabsTabViewModel(DialogService dialogService, PreferencesService preferences) {
-        this.dialogService = dialogService;
-        this.preferences = preferences;
-        this.entryEditorPreferences = preferences.getEntryEditorPreferences();
-    }
-
-    @Override
-    public void setValues() {
-        setFields(entryEditorPreferences.getEntryEditorTabList());
-    }
-
-    public void resetToDefaults() {
-        setFields(preferences.getDefaultTabNamesAndFields());
-    }
-
-    private void setFields(Map<String, Set<Field>> tabNamesAndFields) {
-        StringBuilder sb = new StringBuilder();
-
-        // Fill with customized vars
-        for (Map.Entry<String, Set<Field>> tab : tabNamesAndFields.entrySet()) {
-            sb.append(tab.getKey());
-            sb.append(':');
-            sb.append(FieldFactory.serializeFieldsList(tab.getValue()));
-            sb.append('\n');
-        }
-        fieldsProperty.set(sb.toString());
-    }
-
-    @Override
-    public void storeSettings() {
-        Map<String, Set<Field>> customTabsMap = new LinkedHashMap<>();
-        String[] lines = fieldsProperty.get().split("\n");
-
-        for (String line : lines) {
-            String[] parts = line.split(":");
-            if (parts.length != 2) {
-                dialogService.showInformationDialogAndWait(
-                        Localization.lang("Error"),
-                        Localization.lang("Each line must be of the following form: 'tab:field1;field2;...;fieldN'."));
-                return;
-            }
-
-            // Use literal string of unwanted characters specified below as opposed to exporting characters
-            // from preferences because the list of allowable characters in this particular differs
-            // i.e. ';' character is allowed in this window, but it's on the list of unwanted chars in preferences
-            String unwantedChars = "#{}()~,^&-\"'`ʹ\\";
-            String testString = CitationKeyGenerator.cleanKey(parts[1], unwantedChars);
-            if (!testString.equals(parts[1])) {
-                dialogService.showInformationDialogAndWait(
-                        Localization.lang("Error"),
-                        Localization.lang("Field names are not allowed to contain white spaces or certain characters (%0).",
-                                "# { } ( ) ~ , ^ & - \" ' ` ʹ \\"));
-                return;
-            }
-
-            customTabsMap.put(parts[0], FieldFactory.parseFieldList(parts[1]));
-        }
-
-        entryEditorPreferences.setEntryEditorTabList(customTabsMap);
-    }
-
-    public StringProperty fieldsProperty() {
-        return fieldsProperty;
-    }
-}
